delivery_drone_codebase/requirements.txt.txt
py_trees
dronekit
pymavlink
numpy
pyserial
opencv-python